+++ test/test_restconf.sh
@@ -14,6 +14,7 @@
 # (4) local/backend config. Evhtp only
 # - The tests runs through both (if compiled with evhtp)
 # See also test_restconf2.sh
+# See test_restconf_rpc.sh for cases when CLICON_BACKEND_RESTCONF_PROCESS is set
 
 # Magic line must be first in script (see README.md)
 s="$_" ; . ./lib.sh || if [ "$s" = $0 ]; then exit 0; else return 0; fi